% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calculate_barycenters}
\alias{calculate_barycenters}
\title{Calculate Barycenters of Clusters}
\usage{
calculate_barycenters(data)
}
\arguments{
\item{data}{A dataframe that has been grouped into clusters with 'Angle' and 'Intensity' values.}
}
\value{
A dataframe summarizing the mean 'Angle' and 'Intensity' for each cluster.
}
\description{
This function calculates the barycenters for each cluster based on 'Angle' and 'Intensity' measurements using median values of standardised data.
}
\examples{
\dontrun{
barycenters <- calculate_barycenters(clustered_data)
}
}
\keyword{internal}
