% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{ClusteringFeatureCells}
\alias{ClusteringFeatureCells}
\title{Cluster Feature Cells}
\usage{
ClusteringFeatureCells(
  res_tockyrf,
  percentile = 0.75,
  eps = 4,
  minPts = 2,
  colors = NULL,
  xlim = NULL,
  ylim = NULL,
  test = TRUE
)
}
\arguments{
\item{res_tockyrf}{A list object output from TockyKmeansRF.}

\item{percentile}{A numeric value defining the cutoff for filtering data based on importance scores.}

\item{eps}{The epsilon parameter for DBSCAN, controlling the maximum distance between points in a cluster.}

\item{minPts}{The minimum number of points required to form a dense region in DBSCAN.}

\item{colors}{Optional. A vector to define the colour code for clusters.}

\item{xlim}{Optional. The x range of Timer Blue-Red plot}

\item{ylim}{Optional. The y range of Timer Blue-Red plot}

\item{test}{Logical. The default is TRUE, which enables analysis of test data (recommended).
Optionally, test = FALSE allows analysis of training data.}
}
\value{
Prints plots directly and may return statistical test results if needed.
}
\description{
This function performs density-based clustering on a subset of important data points
}
\examples{
\dontrun{
  clusteringFeatureCells(TockyData, result, percentile = 0.5)
}
}
