% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{match_metaclusters}
\alias{match_metaclusters}
\title{Match Metaclusters Between Training and Testing Sets}
\usage{
match_metaclusters(barycenters_train, barycenters_test)
}
\arguments{
\item{barycenters_train}{A dataframe containing the centroids of clusters in the training set.
Each row represents a cluster with columns `angle_median` and `intensity_median`,
which are the median values of angle and intensity of the cluster, respectively.}

\item{barycenters_test}{A dataframe similar to `barycenters_train` but for the testing set.}
}
\value{
A numeric vector where each element is the index of the training set cluster that
matches with the corresponding test set cluster by the minimal Euclidean distance.
}
\description{
This function matches clusters from the training dataset to clusters in the testing dataset
based on the minimal Euclidean distance between their median-based centroids. Each cluster from the test set
is matched with the closest cluster from the training set, ensuring unique matches. The matching minimizes
the total distance between paired clusters across datasets.
}
\examples{
\dontrun{
barycenters_train <- data.frame(angle_median = c(10, 20, 30, 40),
intensity_median = c(100, 200, 300, 400))
barycenters_test <- data.frame(angle_median = c(9, 21, 31, 39),
intensity_median = c(110, 210, 310, 390))
matches <- match_metaclusters(barycenters_train, barycenters_test)
}
}
\keyword{internal}
