% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{partitionTockyPrepData}
\alias{partitionTockyPrepData}
\title{Split a TockyPrepData object into training and testing sets.}
\usage{
partitionTockyPrepData(x, p = 0.7)
}
\arguments{
\item{x}{A TockyPrepData object.}

\item{p}{A numeric value between 0 and 1 specifying the proportion of data to use for the training set. Defaults to 0.7.}
}
\value{
A list containing the following elements:
  * `trainData`: A TockyPrepData object containing the training data.
  * `testData`: A TockyPrepData object containing the testing data.
}
\description{
Split a TockyPrepData object into training and testing sets.
}
\examples{
\dontrun{
out <- partitionTockyPrepData(my_data)
train_data <- out$trainData
test_data <- out$testData
}

}
