% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{plotImportanceScores}
\alias{plotImportanceScores}
\title{Plot Importance Scores on Test Data}
\usage{
plotImportanceScores(
  res_tockyrf,
  percentile = 0.9,
  test = TRUE,
  plot_mode = "Angle",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{res_tockyrf}{A list object output from TockyKmeansRF.}

\item{percentile}{Numeric, percentile threshold for importance scores (between 0 and 1).}

\item{test}{Logical. If TRUE, Importance Score plot using test data used for creating the TockyKmeansRF model will be used. If FALSE, the training data will be used instead.}

\item{plot_mode}{Either "raw_Timer" for Blue vs Red plots, or "Angle" for Angle vs Intensity plots.}

\item{xlim}{Optional to determine the x ranges of plot. Effective for raw Timer plot only.}

\item{ylim}{Optional to determine the y ranges of plot. Effective for raw Timer plot only.}
}
\value{
A plot of Angle vs Intensity coloured by importance and a colour bar indicating the
        importance scores.
}
\description{
This function plots the importance scores in test data, showing the Angle versus Intensity
coloured by the feature importance derived from a RandomForest model, alongside a colour bar
representing the importance score scale.
}
\examples{
\dontrun{
  # Assuming 'res_tockyrf' is already available from using TockyKmeansRF
  plotImportanceScores(res_tockyrf)
}

}
