% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{plotTockyKmeansRF}
\alias{plotTockyKmeansRF}
\title{Plot ROC Curve from TockyKmeansRF Results}
\usage{
plotTockyKmeansRF(
  res_tockyrf,
  expr_group = NULL,
  ctrl_group = NULL,
  mode = "ROC"
)
}
\arguments{
\item{res_tockyrf}{The output object from `TockyKmeansRF`}

\item{expr_group}{The name of the experimental group within `sampledef`.}

\item{ctrl_group}{The name of the control group within `sampledef`.}

\item{mode}{A character string specifying the type of plot to generate. Valid options are
'ROC' for a Receiver Operating Characteristic curve, and 'PR'
for a Precision-Recall curve. Defaults to 'ROC'.}
}
\value{
A ggplot object representing the ROC curve with the area shaded for the 95% confidence
        interval around the curve.
}
\description{
This function generates a Receiver Operating Characteristic (ROC) curve for the results
obtained from a TockyKmeansRF function, specifically designed for two-group comparisons.
It calculates the area under the curve (AUC) and its confidence interval, and returns
a ggplot object depicting the ROC curve with confidence bands.
}
\examples{
\dontrun{
p <- plotTockyKmeansRF(res_tockyrf)
}

}
