% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Test if cells are inside a polygon gate}
\usage{
point_in_polygon(x, y, poly_coords)
}
\arguments{
\item{x}{The x coordinate of a test point.}

\item{y}{The y coordinate of a test point.}

\item{poly_coords}{A matrix that has two columns and defines a polygon.}
}
\value{
A data frame containing expression data. Note that all inherited values for Gating or Tocky object will not be included.
}
\description{
Test if cells are inside a polygon gate
}
\examples{
\dontrun{
 gate <- locator(type='l', col=2)
 x.gate <- c(gate[[1]], gate[[1]][1])
 y.gate <- c(gate[[2]], gate[[2]][1])
 poly_coords = cbind(x.gate, y.gate)
cd4_data <- point_in_polygon(x, y, poly_coords)
}
}
\keyword{internal}
