% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{violinClusterMFIPlot}
\alias{violinClusterMFIPlot}
\title{Generate a boxplot of MFI (median fluorescence intensity) for each cluster
identified by TockyKmeans.
Generate a boxplot of MFI (median fluorescence intensity) for TockyKmeansRF Feature Clusters}
\usage{
violinClusterMFIPlot(
  x,
  res_tockyrf,
  group = NULL,
  variables = NULL,
  min_cells = 10,
  Timer_positive = FALSE
)
}
\arguments{
\item{x}{A `TockyPrepData` object containing the original flow cytometry data.}

\item{res_tockyrf}{A list object output from `TockyKmeansRF`, which has been further processed
using `ClusteringFeatureCells`.}

\item{group}{Optional. A charaacter vector with the length one when the corresponding group only should be plotted. The default is NULL, which option uses all samples and groups.}

\item{variables}{Optional. A charaacter vector to specify which variables are to be visualised.}

\item{min_cells}{Numeric. The minimum nunmber of cells within a cluster to be analysed. The default is 10.}

\item{Timer_positive}{Logical. Whether to remove Timer negative cells.}
}
\description{
This function visualises marker MFI offeature clusters, other Timer+ cells, and Timer negative cells
}
\examples{
\dontrun{
  data <- data.frame(Angle = runif(100), Intensity = runif(100))
  cell_cluster_id <- dbscan(data, eps = 0.1, minPts = 5)$cluster
  plotHullsGating(data, cell_cluster_id)
}
}
