% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TockyRF.R
\name{violinPlotFeatureCells}
\alias{violinPlotFeatureCells}
\title{Generate Plots for Analysing Cluster Abundance}
\usage{
violinPlotFeatureCells(
  res_tockyrf,
  p_adjust_method = "BH",
  ncol = 3,
  min_cells = 10,
  scatter_plot = FALSE
)
}
\arguments{
\item{res_tockyrf}{A list object output from `TockyKmeansRF`, which has been further processed
using `ClusteringFeatureCells`.}

\item{p_adjust_method}{A method for p-value adjustment in multiple testing using Mann Whitney.
clusteringFeatureCells cen be used.}

\item{ncol}{Number of columns in output figure panel.}

\item{min_cells}{Numeric. The minimum nunmber of cells within a cluster to be analysed. The default is 10.}

\item{scatter_plot}{Logical. If TRUE, scatter plot for Angle and Intensity is generated.}
}
\description{
This function processes clustering results, plots each cluster, and overlays each cluster's convex hull.
It is adaptable to any number of cell_cluster_id.
}
\examples{
\dontrun{
  data <- data.frame(Angle = runif(100), Intensity = runif(100))
  cell_cluster_id <- dbscan(data, eps = 0.1, minPts = 5)$cluster
  violinPlotFeatureCells(data, cell_cluster_id)
}
}
